PointFT/requirements.txt
h5py==3.6.0
matplotlib==3.4.3
munch==2.5.0
open3d==0.13.0   
torch==1.9.0
pyyaml==5.4.1